move colors to a new "module"
