+++ modules/colors/colors.js
@@ -1,0 +1,141 @@
+// @flow
+
+export const aqua = '#7FDBFF'
+export const black = '#111111'
+export const blue = '#0074D9'
+export const fuchsia = '#F012BE'
+export const gray = '#AAAAAA'
+export const green = '#2ECC40'
+// export const lime = '#01FF70'
+// export const maroon = '#85144b'
+export const navy = '#001f3f'
+// export const olive = '#3D9970'
+export const orange = '#FF851B'
+export const purple = '#B10DC9'
+export const red = '#FF4136'
+export const silver = '#DDDDDD'
+// export const teal = '#39CCCC'
+export const white = '#FFFFFF'
+export const yellow = '#FFDC00'
+export const transparent = 'transparent'
+
+export const semitransparentGray = 'rgba(0,0,0,0.2)'
+
+// MARK: System Colors
+export const infoBlue = 'rgb(47, 112, 225)'
+export const success = 'rgb(83, 215, 106)'
+export const warning = 'rgb(221, 170, 59)'
+export const danger = 'rgb(229, 0, 15)'
+
+// MARK: Whites
+export const antiqueWhite = 'rgb(250, 235, 215)'
+export const oldLace = 'rgb(253, 245, 230)'
+export const ivory = 'rgb(255, 255, 240)'
+export const seashell = 'rgb(255, 245, 238)'
+export const ghostWhite = 'rgb(248, 248, 255)'
+export const snow = 'rgb(255, 250, 250)'
+export const linen = 'rgb(250, 240, 230)'
+
+// MARK: Grays
+export const black25Percent = 'rgb(63.75, 63.75, 63.75)'
+export const black50Percent = 'rgb(127.5, 127.5, 127.5)'
+export const black75Percent = 'rgb(191.25, 191.25, 191.25)'
+export const warmGray = 'rgb(133, 117, 112)'
+export const coolGray = 'rgb(118, 122, 133)'
+export const charcoal = 'rgb(34, 34, 34)'
+
+// MARK: Blues
+export const teal = 'rgb(28, 160, 170)'
+export const steelBlue = 'rgb(103, 153, 170)'
+export const robinEgg = 'rgb(141, 218, 247)'
+export const pastelBlue = 'rgb(99, 161, 247)'
+export const turquoise = 'rgb(112, 219, 219)'
+export const skyBlue = 'rgb(0, 178, 238)'
+export const indigo = 'rgb(13, 79, 139)'
+export const denim = 'rgb(67, 114, 170)'
+export const blueberry = 'rgb(89, 113, 173)'
+export const cornflower = 'rgb(100, 149, 237)'
+export const babyBlue = 'rgb(190, 220, 230)'
+export const midnightBlue = 'rgb(13, 26, 35)'
+export const fadedBlue = 'rgb(23, 137, 155)'
+export const iceberg = 'rgb(200, 213, 219)'
+export const wave = 'rgb(102, 169, 251)'
+
+// MARK: Greens
+export const emerald = 'rgb(1, 152, 117)'
+export const grass = 'rgb(99, 214, 74)'
+export const pastelGreen = 'rgb(126, 242, 124)'
+export const seafoam = 'rgb(77, 226, 140)'
+export const paleGreen = 'rgb(176, 226, 172)'
+export const cactusGreen = 'rgb(99, 111, 87)'
+export const chartreuse = 'rgb(69, 139, 0)'
+export const hollyGreen = 'rgb(32, 87, 14)'
+export const olive = 'rgb(91, 114, 34)'
+export const oliveDrab = 'rgb(107, 142, 35)'
+export const moneyGreen = 'rgb(134, 198, 124)'
+export const honeydew = 'rgb(216, 255, 231)'
+export const lime = 'rgb(56, 237, 56)'
+export const cardTable = 'rgb(87, 121, 107)'
+
+// MARK: Reds
+export const salmon = 'rgb(233, 87, 95)'
+export const brickRed = 'rgb(151, 27, 16)'
+export const easterPink = 'rgb(241, 167, 162)'
+export const grapefruit = 'rgb(228, 31, 54)'
+export const pink = 'rgb(255, 95, 154)'
+export const indianRed = 'rgb(205, 92, 92)'
+export const strawberry = 'rgb(190, 38, 37)'
+export const coral = 'rgb(240, 128, 128)'
+export const maroon = 'rgb(80, 4, 28)'
+export const watermelon = 'rgb(242, 71, 63)'
+export const tomato = 'rgb(255, 99, 71)'
+export const pinkLipstick = 'rgb(255, 105, 180)'
+export const paleRose = 'rgb(255, 228, 225)'
+export const crimson = 'rgb(187, 18, 36)'
+
+// MARK: Purples
+export const eggplant = 'rgb(105, 5, 98)'
+export const pastelPurple = 'rgb(207, 100, 235)'
+export const palePurple = 'rgb(229, 180, 235)'
+export const coolPurple = 'rgb(140, 93, 228)'
+export const violet = 'rgb(191, 95, 255)'
+export const plum = 'rgb(139, 102, 139)'
+export const lavender = 'rgb(204, 153, 204)'
+export const raspberry = 'rgb(135, 38, 87)'
+export const fuschia = 'rgb(255, 20, 147)'
+export const grape = 'rgb(54, 11, 88)'
+export const periwinkle = 'rgb(135, 159, 237)'
+export const orchid = 'rgb(218, 112, 214)'
+
+// MARK: Yellows
+export const goldenrod = 'rgb(215, 170, 51)'
+export const yellowGreen = 'rgb(192, 242, 39)'
+export const banana = 'rgb(229, 227, 58)'
+export const mustard = 'rgb(205, 171, 45)'
+export const buttermilk = 'rgb(254, 241, 181)'
+export const gold = 'rgb(139, 117, 18)'
+export const cream = 'rgb(240, 226, 187)'
+export const lightCream = 'rgb(240, 238, 215)'
+export const wheat = 'rgb(240, 238, 215)'
+export const beige = 'rgb(245, 245, 220)'
+
+// MARK: Oranges
+export const peach = 'rgb(242, 187, 97)'
+export const burntOrange = 'rgb(184, 102, 37)'
+export const pastelOrange = 'rgb(248, 197, 143)'
+export const cantaloupe = 'rgb(250, 154, 79)'
+export const carrot = 'rgb(237, 145, 33)'
+export const mandarin = 'rgb(247, 145, 55)'
+
+// MARK: Browns
+export const chiliPowder = 'rgb(199, 63, 23)'
+export const burntSienna = 'rgb(138, 54, 15)'
+export const chocolate = 'rgb(94, 38, 5)'
+export const coffee = 'rgb(141, 60, 15)'
+export const cinnamon = 'rgb(123, 63, 9)'
+export const almond = 'rgb(196, 142, 72)'
+export const eggshell = 'rgb(252, 230, 201)'
+export const sand = 'rgb(222, 182, 151)'
+export const mud = 'rgb(70, 45, 29)'
+export const sienna = 'rgb(160, 82, 45)'
+export const dust = 'rgb(236, 214, 197)'
