+++ modules/colors/gradients.js
@@ -1,0 +1,18 @@
+// @flow
+
+// MARK: gradients
+export const redToPurple = ['rgb(249,64,77)', 'rgb(217,37,111)']
+export const orangeToRed = ['rgb(250,120,37)', 'rgb(242,38,83)']
+export const yellowToGoldDark = ['rgb(234,185,29)', 'rgb(226,123,19)']
+export const yellowToGoldLight = ['rgb(238,209,0)', 'rgb(237,158,0)']
+export const grassToLime = ['rgb(125,209,24)', 'rgb(88,186,34)']
+export const seafoamToGrass = ['rgb(15,201,121)', 'rgb(58,181,42)']
+export const tealToSeafoam = ['rgb(3,220,156)', 'rgb(19,188,156)']
+export const lightBlueToBlueLight = ['rgb(1,194,199)', 'rgb(10,158,196)']
+export const lightBlueToBlueDark = ['rgb(5,187,219)', 'rgb(25,145,234)']
+export const darkBlueToIndigo = ['rgb(48,112,228)', 'rgb(95,72,229)']
+export const purpleToIndigo = ['rgb(129,88,252)', 'rgb(92,80,236)']
+export const magentaToPurple = ['rgb(213,69,196)', 'rgb(160,50,230)']
+export const pinkToHotpink = ['rgb(248,102,149)', 'rgb(209,44,159)']
+export const grayToDarkGray = ['rgb(118,134,157)', 'rgb(85,96,124)']
+export const navyToNavy = ['rgb(32,67,107)', 'rgb(0,48,104)']
