+++ source/lib/colors.js
@@ -1,0 +1,22 @@
+// @flow
+
+// MARK: St. Olaf colors
+export const sto = {
+	black: '#242121',
+	darkGray: '#53585A',
+	mediumGray: '#E0DAD6',
+	lightGray: '#F7F5F2',
+	gold: '#E3A025',
+	lightGold: '#F7C55A',
+	cranberry: '#B11556',
+	lime: '#A3D65C',
+	orange: '#FA8D29',
+	red: '#E84A36',
+	lightBlue: '#83DBE0',
+	blue: '#00ADD9',
+	teal: '#00827E',
+	purple: '#6E3A5D',
+	navy: '#233746',
+	mint: '#CFE1D7',
+	white: '#FFFFFF',
+}
