+++ source/views/home/button.js
@@ -6,7 +6,8 @@ import Icon from 'react-native-vector-icons/Entypo'
 import type {ViewType} from '../views'
 import LinearGradient from 'react-native-linear-gradient'
 import {Touchable} from '../../components/touchable'
-import * as c from '../../components/colors'
+import {transparent} from '@frogpond/colors'
+import {homescreenForegroundDark, homescreenForegroundLight} from './colors'
 import {iPhoneX} from '../../components/device'
 
 type Props = {
@@ -108,17 +109,17 @@ const styles = StyleSheet.create({
 		paddingHorizontal: cellHorizontalPadding,
 	},
 	icon: {
-		backgroundColor: c.transparent,
+		backgroundColor: transparent,
 	},
 	text: {
-		backgroundColor: c.transparent,
+		backgroundColor: transparent,
 		fontFamily: Platform.OS === 'ios' ? 'System' : 'sans-serif-condensed',
 		fontSize: 14,
 	},
 	lightForeground: {
-		color: c.homescreenForegroundLight,
+		color: homescreenForegroundLight,
 	},
 	darkForeground: {
-		color: c.homescreenForegroundDark,
+		color: homescreenForegroundDark,
 	},
 })
