+++ source/views/home/colors.js
@@ -1,0 +1,4 @@
+// @flow
+
+export const homescreenForegroundLight = 'rgba(255, 255, 255, 0.9)'
+export const homescreenForegroundDark = 'rgba(0, 0, 0, 0.65)'
